+++ addons/ofxAssimpModelLoader/src/ofxAssimpModelLoader.cpp
@@ -14,6 +14,17 @@ ofxAssimpModelLoader::~ofxAssimpModelLoader(){
     clear();
 }
 
+ofxAssimpModelLoader::ofxAssimpModelLoader(const ofxAssimpModelLoader& orig) {
+	scene = NULL;
+	clear();
+
+	if(orig.file.exists()) {
+		loadModel(orig.file.getAbsolutePath());
+	} else if(orig.scene != NULL) {
+		ofLogWarning("ofxAssimpModelLoader") << "can't copy loader without a file reference";
+	}
+}
+
 //------------------------------------------
 bool ofxAssimpModelLoader::loadModel(string modelName, bool optimize){